Support Pldmd Stand-Alone build using meson
This commit is for pushing the work done so far in making pldmd
as a standalone build using meson as build system.
The meson framework for pldmd was decided to have pldmd
as standalone build which would be useful for running
independent test cases for pldmd.

The code base for modules may be older than the current level.
The commit is still a work in progress.

Tested:
  WIP

Signed-off-by: Sabarish Nair <sabarishx.nair@intel.com>